exports: Fix module definition file for MinGW/MSYS
GNU linkers expect it to be passed as an object file. This is
described in the ld man page [1]:

> In addition, the linker fully supports the standard "*.def" files,
> which may be specified on the linker command line like an object
> file (in fact, it should precede archives it exports symbols from,
> to ensure that they get linked in, just like a normal object file).

I've verified this to be the case in the MSYS2 CLANG64, UCRT64 and
MINGW32 flavors.

Alternatively, one could modify the version script generator to output a
GNU-format version script if AOM_TARGET_SYSTEM=Windows and
AOM_MSVC=FALSE, but this was the smaller approach.

[1]: https://linux.die.net/man/1/ld

Change-Id: I51d24fb623419647929470474fee591579795b57
(cherry picked from commit a5e0c896cdb235476513a70e17d322583a51c9df)