cmake: Fix build with Clang/CL
It is possible to use the Clang/CL compiler when targeting MSVC;
unfortunately, this flavour requires using the same flags as GCC to
handle intrinsics codegen; it won't enable SSE4.1 by default like MSVC.

For the GNU flavour, an additional fix is needed so that C++14 is used
and the glibc math library isn't linked against.

This fixes compiling aom with Clang/CL 15.0.1 (MSVC 17.6 Preview 3).

Change-Id: I54140a1645f539944657c2b18dd801b57e422216