cmake: Fix build with Clang/CL
It is possible to use the Clang/CL compiler when targeting MSVC;
unfortunately, not only this flavour requires using the same flags
as GCC to handle intrinsics codegen, the default is solely SSE2
(like with stock Clang, and unlike MSVC's own default of SSE4.1 [1]).
The MSVC STL also requires C++14. For this reason, when the GNU
flavour is detected, we apply the correct flag style and use C++14
as the C++ standard.

An additional fix is also needed so that the glibc math library
isn't linked against when using this kind of Clang compiler.

This fixes compiling aom with Clang/CL 15.0.1 (MSVC 17.6 Preview 3).

[1]: https://learn.microsoft.com/en-us/cpp/build/reference/arch-x64
for the x86_64 case.

Change-Id: I54140a1645f539944657c2b18dd801b57e422216
(cherry picked from commit 9763f658502e19b352ec2ae66a06b5bf6491e50b)