Add Neon implementations for remaining h predictors
We already have Neon implementations of the h predictors for all square
block sizes except 64x64, so add the remaining cases and update
tests/speed to match. We also clean up and refactor the existing cases
to expose some helper functions we can reuse in the new implementations.

On Neoverse N1 these new implementations are worth about a 33% geomean
performance improvement over the C versions on both Clang 15 and GCC 12.
On Neoverse V1 there is a smaller improvement, with a geomean
improvement of around 3% on Clang 15 and 15% on GCC 12.

Change-Id: Ib9dd1e2c924c41d620635b8e3acb813242642cb3