Corner match: Use bidirectional matching
Remember the best match for each feature in both the source and ref
frames. Only generate correspondences for points which are each other's
best matches. This removes a lot of spurious correspondences, which
helps improve the average quality of the generated correspondences.

In addition, rearrange the code to be much more efficient. If there
are n features in each of the source and ref frames, then the matching
loop is run O(n^2) times. Therefore it is beneficial to push as much
work as possible out into pre- and post-processing loops which run
only O(n) times.

Results @ "good" mode speed 4:

     Compared to      | BDRATE-PSNR | BDRATE-SSIM | Encode time
----------------------+-------------+-------------+-------------
Previous corner match |   -0.037%   |   -0.051%   |   +1.955%
       Disflow        |   +0.010%   |   -0.006%   |  +14.189%

No change to encoder output, as we currently use disflow.

Change-Id: I0abd73717925623eeb3948889eae3ca38c61811d