Fix comp_mask_variance speed tests
Speed tests were run with all possible block sizes, including those
whose wedge parameters are NULL, triggering a segmentation fault. This
was not a problem for regular tests as they are run within a loop that
is simply not entered when there are no wedge paramters. All
comp_mask_variance tests are now run only on relevant block sizes.

Change-Id: I6b801ab26293daac28481596c87677388f450fc3