Remove extend_minq_fast.
Remove feature in VBR to aggressively lower the minimum active Q

This feature was designed to try and help with extreme overshoot but
in the current code is not consistently helping either with rate
accuracy or quality. It does however add to complexity of the code.

A subsequent patch will greatly improve rate accuracy on most clips
for both the undershoot and overshoot cases.

Small mixed impact on metrics.

STATS_CHANGED

Change-Id: I231336c632eb210c346bc3e1acd428160e3dd000