Improve rate control in vbr.
Alter the calculation and use of extend_minq and extend_maxq

This patch changes the basis for reacting to rate error to a more local
measure though the overall rate situation is also taken into account.
It also changes the extent and rate at which the active minimum
and maximum q can be adjusted and allows for minq to be adjusted
both up and down.

On average the metrics performance of the test sets is improved and
for the majority of clips rate accuracy is also better. Metrics gains
are most consistent and largest when a low rate error threshold is
set (high accuracy required).

Rate thresholds +/- 10%
lowres	-1.747	-2.712	-1.431
lowres2	-0.556	-1.195	-0.623
midres	-1.449	-1.982	-1.498
midres2	-1.058	-1.484	-1.218
ugc360P -0.540	-0.766	 0.083
hdres	-3.509	-4.584	-3.991

Rate thresholds +/- 25%
lowres	-0.825	-1.092	-0.533
lowres2	 0.148	 0.223	 0.126
midres	-0.209	 0.205	-0.231
midres2	 0.347	 0.677	 0.308
ugc360p	-0.614	 0.012	-0.176
hdres	-2.003	-2.254	-2.354

There is a significantly lower incidence of cases where there is a large
rate overshoot, though for some of these clips this is accompanied by
weaker metrics performance.

On some low motion content there is more undershoot especially at low
target rates as the maximum amount by which minq can be lowered and
the rate at which it can be lowered has been reduced. However, many of
these low motion clips (for example mobile calender and jets in the
hdres set and lyrics_360_cb20... in the ugc set are also amongst the
biggest winners in metrics terms with gains of 10-20%).

Change-Id: Ia0c222297e4e567303e9dac3b786d5d98b450eef