Disable Loop restoration for multi-thread case
In speeds 5 and 6, enabling of Loop restoration in multi-thread
case results in a higher encode time increase as compared to
single thread. Hence, disabling the same for multi-thread case.
Note that the output of single-thread and multi-thread will not
match after this CL for speed 5 and 6.

STATS_CHANGED for speed 5, 6 with multi-threading enabled

Change-Id: I13485b3d48eb6580f8818329496c8cb3613da950