Document invalid OP index for target-seq-level-idx
Document the current behavior of AV1E_SET_TARGET_SEQ_LEVEL_IDX and
target-seq-level-idx when an invalid operating point index is specified
in the value.

The current behavior is a no-op and successful return, and is checked by
a test in test/level_test.cc. Although I think it would be better to
return AOM_CODEC_INVALID_PARAM in this case, I did not change the
behavior because I am not sure if the current behavior is intended.

Change-Id: Ib35d1a2e73ed5785edc715868b60d1dfe1770f56