Use TBL for z2 predictor on AArch64 Neon
The existing code loads each lane of the resulting vector independently,
however in this case the range of possible values for indices is
sufficiently small such that we can simply load all possible values and
then use TBL to select the appropriate ones.

The 'q' variants of the TBL intrinsics are only available in AArch64, so
restrict the new implementation appropriately.

Change-Id: I9d168a093c91525aed7a60e3ef7bd877c293b43e