Refactor tests for aom_convolve8 functions
Currently it is not possible to test both aom_convolve8 and
aom_highbd_convolve8 functions when CONFIG_AV1_HIGHBITDEPTH is enabled,
which is impractical. This patch splits the ConvolveTest test suite into
LowbdConvolveTest and HighbdConvolveTest, so that both bitdepths can be
tested on a build with CONFIG_AV1_HIGHBITDEPTH enabled.

Change-Id: Id44e883f297bebb37e8708eda595e0aae9f5031b