Fix sync_enc_workers to read error status of master thread
In sync_enc_workers(), winterface->sync() is invoked on the
worker threads and the same is used to read the error status of
each thread. The sync of master thread was avoided to facilitate
FPMT, but its error status still needs to be read to facilitates
the error propagation. This CL initializes 'had_error' with the
error status of the master thread to fix this issue.

Bug: aomedia:3276

Change-Id: I0a4acb305cdd6ef38a7d2848440aa64156f16175