Fix sync_fpmt_workers() to get correct error information
In case of FPMT, worker->data1 will point to the respective
frames cpi, but the same will be overwritten by the thread
specific argument while executing the low-level multi-threaded
modules for the same frame. Due to this, the 'error' information
is being dereferenced from the incorrect cpi in
sync_fpmt_workers(). This CL fixes this issue by using the
correct cpi instance from ppi.

Bug: aomedia:3276

Change-Id: I3f9a2f687d02c5740447a8c97b16e87380d954e9