av1_fwd/inv_txfm1d_test.cc: Standardize var names
The original motivation for this CL is to fix a -Wshadow warning in
av1_fwd_txfm1d_test.cc. While choosing a new name for the loop index
variable, I decided to use use the same name that is used in similar
loops elsewhere. In libaom and libvpx, the most common index variable
name used with `count_test_block` is `i`, so I standardize on `i`.

In these two files, standardize on the variable name `txfm_size` for an
element of the txfm_size_ls array, and standardize on the index variable
name `ni` with `txfm_size`.

Bug: aomedia:3450
Change-Id: I98ea71a2f2014a7507af6e5ae8cf6015c3fc23a8