rtc: Add control to set target bitrate
For 1 pass CBR, and only for single layer for now.

This allows to bypass a lot of code via aom_codec_enc_config_set.
For RTC applications that often update the encoder with new
target bitrate.

Tested manually with the code snippet in svc_encoder_rtc.
This exposed an incorrect reset for RTC in change_config():
https://aomedia-review.googlesource.com/c/aom/+/176202

Change-Id: I5635f4c15952b933158215734d555853c7c69b1a