Add Armv8.4 SDOT Neon implementation of aom_convolve8_horiz
Add an Armv8.4 Neon SDOT (dot-product) path for aom_convolve8_horiz.
The existing MLA-based implementation of is retained and used on
target CPUs that do not implement the SDOT instruction (or CPUs
executing in AArch32 mode.) The availability of the SDOT instruction
is indicated by the feature macro __ARM_FEATURE_DOTPROD.

An Armv8.4 SDOT Neon implementation of aom_convolve8_vert will be
added in a subsequent patch.

The Neon code added in this change is mostly a port of the Armv8.4
Neon implementation of the 'vpx_convolve8_horiz_neon' function in
libvpx.[1]

[1] https://github.com/webmproject/libvpx/blob/main/vpx_dsp/arm/vpx_convolve8_neon.c

Change-Id: Ia430ac4652f20908bf1b721842b4f243da99155b