rtc: Fix to last_source when frame_dropper is enabled.
last_source is used for scene detection and superblock
content analysis. last_source is meant to be the source
corresponding to the LAST reference, as this is mainly
used for prediction. So when frames are dropped we need
to make sure the last_source corresponds to the LAST
reference buffer (which for single layer is the last
encoded frame).

For temporal layers, where LAST reference is not the previous
encoded frame (2/4 behind for 2/3 temporal layers), this
issue has been fixed via the logic of av1_svc_set_last_source()
and the additional buffer source_last_TL0.

Reuse this logic for the case where the frame_dropper is
enabled.

This improves the quality (significant reduction in artifacts)
for screen content when frames are dropped (due to scrolling or
slide changes).

The current fix mainly applies to single layer. There may be
follow-up fix needed for the case of frame_dropper and SVC.

Change-Id: I774c7d3fc9f0b07c2438aadd08c24291a75d25b2
(cherry picked from commit 6154a1ed4dd925b05be9fe83850ad5294d2de9ab)