Make the screen content decision early
Currently, the screen content decision is made after the temporal
filtering and TPL. This makes the logic complicated if we need to
optimize temporal filtering and TPL differently for screen and
non-screen contents. The screen content decision is independent,
which is based on source key frame information. Therefore, we
move this decision making before temporal filtering and TPL to
keep the logic simple.

Change-Id: I5d83ea6b203c311c37cc2ff1cf59466c1c5dd2e0