rtc: Fix to scene change detection, for drop_frame
When a frame is dropped: in the postencode_drop reset
the running average source_sad (avg_source_sad) to 0,
on detected scene changes (high_source_sad = 1).
This is because scene detection is based on current frame
source_sad relative to the running average, and this can
suddenly be large (on a scene change frame that is dropped)
and then missed detection occurs (scene change won't be set
for subsequent frame).

Also remove 1 sb from bottom border, as content changes
along those borders are not significant and there seems to be
an issue when using source_last_tl0 as last_source.

This improves quality for screen content if frames are dropped
(with --drop_frame > 0).

Change-Id: I66b552a2e5eeefe93aeea9abefd2c9940c1ebe5e
(cherry picked from commit 414a2fcbb1bb0f4ac471613e12004651f4d67c25)