rtc: Fix to scene change detection, for drop_frame
When a frame is dropped: in the postencode_drop we
subtract off the update to the frame source_sad
running average (avg_source_sad) from the current frame
soure_sad (frame_source_sad), on detected scene changes.
(high_source_sad = 1). This is because scene detection is
based on current frame source_sad relative to the running average,
and this can suddenly be large (on a scene change frame that is
dropped) and then missed detection occurs (scene change won't be
set for subsequent frame).

Also remove 1 sb from top and bottom border, as content changes
along those borders are not significant and there seems to be.
an issue when using source_last_tl0 as last_source.

This improves quality for screen content if frames are dropped
(with --drop_frame > 0).

Change-Id: I66b552a2e5eeefe93aeea9abefd2c9940c1ebe5e