Tidy up warp model type checks
Currently, there is some logic inside warp_plane() which fixes up
ROTZOOM models so that the warp matrix has the right properties.
However, if the warp model isn't set up properly by this point,
it's too late to do anything about it, because we will already have
computed the shear parameters from the incorrect warp matrix.
So all this logic can do is silently hide incorrectly computed models.

So replace this logic with a (more thorough) set of assertions at the
start of av1_get_shear_params(). This is a good place for the checks,
as the model must be correct by the time this function is called in
order to get correct (ie, spec-conformant) output.

Lastly, rearrange some encoder-side code to make it clearer where the
relevant constraints are enforced.

No effect on encoder or decoder output, as all models are set up correctly
at the moment.

Change-Id: I32fe59e05b3b011f23c3165da868fdc166817b65