Fix the input to Hadamard functions in unit test
The input to Hadamard transform is the residual block, whose pixels
can range from -((1 << BitDepth) - 1) to (1 << BitDepth) - 1. Earlier
values from -(1 << BitDepth) to (1 << BitDepth) - 1 were  being tested.
As -(1 << BitDepth) is not a valid value, changed the input to take
values in the range [-((1 << BitDepth) - 1), ((1 << BitDepth) - 1)].

The functions Rand9Signed() and Rand13Signed are deleted since they
are not used by any other caller.

Change-Id: I10b358a2ff7168ac595849d3c78a91cfa6b3b24d