Use transpose helper in warp affine vertical_filter_neon
The existing implementation of the vertical filter does a partial
transpose of both the input and filter vectors in order to make use of
the pairwise addition instructions. It appears to be marginally faster
to do the straightforward approach of fully transposing the input data
and then multiplying across, matching what we do in the new horizontal
filter implementation. This approach also allows us to reuse the
existing transpose helper functions.

The old horizontal filter implementation required the filters to be
permuted to avoid intermediate overflow before widening to 32-bit, and
this permute also affected the vertical filter implementation. Since we
are rewriting it anyway, undo this permute in the horizontal filter to
simplify things.

The warp affine speed tests report a roughly 4% reduction in time as a
result of this patch, when built with both Clang 16 and GCC 12 and
meaasured on a Graviton 3 machine.

Change-Id: I5952ef63b15857d50f3b173108cd8a22a0850d23