Refactor av1_warp_affine_neon horizontal/vertical filter loops
This patch simply moves the loops that call into the horizontal and
vertical filters into their own functions. This is designed to allow for
further optimisation of the loop structure in later commits, in
particular to allow for specialising the case where p_width==4 and
therefore half of the filter is unused.

Change-Id: I76851e72643344a9f5149f1d0d027cfaa292c76b