rtc: Allow for increase in palette color_thresh
The current palette prediction is enabled only for
color_threshold <= 64, which was set empirically,
since palette prediction generally works better when
the number of colors/levels in the block is not so large.

But there are cases (in screen real-time mode) where a
larger threshold improves coding efficiency. So here we
allow for larger color_threshold for palette, based on
frame source_sad, block source variance and the UV distortion.
For the latter keep track of color/UV distortion during the
inter mode search.

Also adjust some thresholds on sse and source_variance
for testing palette.

This generally improves quality for screen and removes
some artifacts observed in a few clips.

Add this increase in palette color_threshold as a speed
feature and set it on for speeds 7 - 10, in real-time screen.

Gains for speeds 7-10 on rtc_screen set:
              avg/ovr/ssim     IC% speedup
speed 7:   0.378/-1.533/-0.343,  -0.498
speed 8:  -1.133/-2.740/-1.323   -0.690
speed 9   -0.514/-2.677/-1.636   -0.868
speed 10: -0.987/-3.370/-2.274   -0.639

Change-Id: Ie3819fed2015ac779a4298fdcbe794524f2fc6f6