rtc: Allow for increase in palette color_thresh
For nord_pickmode screen content mode:
allow for larger color_threshold for palette,
based on frame source_sad, block source variance and
the UV distortion. For the latter keep track of color/UV
distortion during the inter mode search.

Also adjust some thresholds on sse and source_variance
for enabling palette.

This generally improves quality for screen and removes
some artifacts observed in a few clips.

Add this increase in palette color_threshold as a speed
feature and set it on for speeds 7 - 10.

Gains for speeds 7-10 on rtc_set:
              avg/ovr/ssim     IC% speedup
speed 7:   0.378/-1.533/-0.343,  -0.498
speed 8:  -1.133/-2.740/-1.323   -0.690
speed 9   -0.514/-2.677/-1.636   -0.868
speed 10: -0.987/-3.370/-2.274   -0.639

Change-Id: Ie3819fed2015ac779a4298fdcbe794524f2fc6f6