Specialise av1_warp_affine_neon vertical filter for p_width == 4
For the small cases we only need to do half the work in the vertical
filter, so specialise this case to make this discarding more obvious to
the compiler.

Benchmarking this on a Neoverse N2 machine with Clang 16 and GCC 12 show
a geomean 4% reduction in runtime for the warp affine speed tests
(mostly benefiting the 4xh cases).

Change-Id: Iadead2ccd4d8ff8a57df479099e74b09150d8341