Specialise av1_warp_affine_neon horizontal filter for p_width == 4
Currently we calculate the full eight-wide filter in the horizontal pass
even if the vertical filter will only use the first four elements of the
result.

We can improve things by checking p_width in the horizontal filter and
only calculating the low four elements of the result. The high half is
set to zero to allow us to use the same uint16x8_t tmp array.

Benchmarking this with the warp affine speed tests on a Neoverse N2
machine with Clang-16 and GCC-12 gives a geomean 9.7% reduction in
runtime, with the p_width == 4 cases reducing by 15-30% and the
remainder at parity.

Change-Id: I616152117e7c238eadbd81825f877e8be4f8a7c4