Remove references to tiles in LR code
During AV1 development, we were unsure whether loop restoration would
be applied per-tile or across the whole frame, and had code to handle
either case. Eventually we settled on applying LR across the whole
frame, but the code was only partially updated to match this decision.

This patch applies some wide-ranging cleanups, now that we know that
loop restoration is always applied across the whole frame, without
any special treatment for tile boundaries.

This also cleans up a long-standing issue where "tile" in the LR code
could refer to either a coding tile, or a loop restoration unit. Now,
all the cases where a coding tile was meant have been removed, and the
rest have been renamed to make it clearer that they mean a loop
restoration unit.

Note, however, that the encoding of loop restoration information is
still done per-tile, even though the processing is independent of
the tile structure.

Change-Id: I3f820e06878d20574f472dad922e1171ee7f6e18