rtc-screen: Reduce blocking artifacts on flat areas
For screen in nonrd_pickmode: for spatially flat
blocks compute the block source_sad and use it to
flag whether the coding block has any motion. This
is then used for decision on skipping inter-modes.

Currently the only flag is at the superblock level.
Doing it at the coding level allows for better mode
testing decision.

This reduces blocking artfiacts for speed >= 7.

Change-Id: I5118ec4175e4f940c8474c7822c464c54487b196