Replace plane_rect with plane_w, plane_h
The value plane_rect, which gets passed around a lot in the loop
restoration code, always has the form {0, plane_w, 0, plane_h}.
So, to reduce space usage a little, and make the code clearer, we
pass around the plane size directly instead of this rectangle.

Also normalize naming - previously the width was referred to by
a mix of "plane_w" and "plane_width", and similarly for height.
Standardize to "plane_w" and "plane_h".

Change-Id: I1c6a84a78a6b1db2c71fcf577d88305958741de2