Add debugging infrastructure for LR costing
There is currently a bug in loop restoration costing when tiles are
in use, caused by the fact that we are calculating the bit cost of
filter parameters based on an incorrect set of reference parameters.

Add debug logic which can detect this bug. The logic is disabled by
default, but can be manually enabled to verify that the bug exists
in this commit, and that the following commit fixes it.

Change-Id: Idcaf33029330b84a3b4d505fce83cc3f8c03db8b