Test aom_get_blk_sse_sum on relevant block sizes only
aom_get_blk_sse_sum is only ever called with a list of specific block
sizes, however the tests were calling it with a range of values of
width, most of which weren't in that list. Use only relevant values for
the block size so that the code doesn't have to accomodate for values
that are never actually used.

Change-Id: I6f25969369939e0418c3ce0b27d56ca67d25c35a