Select loop restoration params in one pass
Switch the loop ordering during restoration search: instead of
evaluating one filter type at a time for all RUs, we now evaluate
all filter types for one RU at a time. This arrangement is nicer
for two reasons:

1) It's more cache friendly - the previous code cycles the image
   through the CPU cache three times, while the new code only
   cycles it through once

2) We can move some information from RestUnitSearchInfo (one per
   restoration unit in the Y plane) to RestSearchCtxt (one, total).
   We do have to duplicate some values in RestSearchCtxt to keep
   the correct behaviour, but overall this reduces memory usage
   on all but the smallest videos.

Along the way, this patch fixes an obscure bug where the
skip_sgr_eval flag in the RestUnitSearchInfo struct was not always
set by search_wiener, and was not reset to 0 between planes. So,
if things lined up in just the right way, it was possible for the
encoder to skip SGR search for a chroma RU based on a stale flag.

Only a handful of videos are affected by this bugfix. Using the
standard AV1 test sets across speeds 1-4, only 8 videos are affected,
and the average BDRATE impact on those videos is -0.021%.

STATS_CHANGED

Change-Id: I235db18814969472decbcacec7111113e98d24ad