Fix warp filter speed tests zero parameters
The existing speed tests are parameterized over
is_{alpha,beta,gamma,delta}_zero however these parameters are ignored by
generate_warped_model. In particular the branch they are used in is
never taken when used with the deterministic RNG that the speed test
relies on.

This commit updates the test to force these parameters to zero as
expected by the benchmarking, and additionally to force non-zeros in the
opposite case. We also extend the print code to allow us to easily
verify that the new speed tests are behaving as expected.

Change-Id: I3387dbfb992dd7834d82fb1a2a36a83cb88ab1fc