Specialize av1_warp_affine_neon for alpha/beta=0
Profiling suggests that 0 is a very common value for these parameters so
it is worth specializing the implementation for them.

We could explicitly rewrite the code to avoid loading filters multiple
times with the same offset or hoist the filter loading out of the loop
in the beta=0 case, but compilers seem happy enough to do the
simplifications for us after inlining horizontal_filter_neon already.

When benchmarking with Clang-16 and GCC-12 on a Neoverse N2 machine we
see a geomean ~9.1% reduction in the time reported by the warp affine
speed tests, with a significant speedup reported in the common
alpha=beta=gamma=delta=0 case.

Change-Id: I0942cd641a5cd34197ae4f1415e8302f00613066