Add Neon implementation of sobel_filter
This is the first patch of a series of patches to add a Neon
implementation of aom_compute_flow_at_point. Since this function only
calls static helpers, most of the scalar C implementations of those
helpers have been copied over to the Neon file. Subsequent patches will
gradually replace them with a Neon implementation.

Add the corresponding tests as well.

Change-Id: I6e289862c52d7585af51eb81b43f2bb35fd47a21