rtc-screen: Set dc_blk_pred_level palette prediction
For nonrd_pickmode (speed >= 7) screen content mode,
on delta frames, set the dc_blk_pred_level to 3.
This is the more aggressive level to predict a block
to be skip or dc only. This will only affect palette
prediction by speeding up the tx_search.

~1-2% speedup on rtc_screen acrosss speeds 7-11,
with neutral/small bdrate change.

Change-Id: I2d58497375b5e5fab17235c071a219792b74548a