Handle Arm/AArch64 runtime feature detection in test_libaom.cc
The tests themselves are not feature-gated in the same way that they are
used in the rest of the codebase since they are not controlled by
rtcd.pl. This means that tests that assume the existence of features not
present on the target can cause SIGILL to be thrown.

This commit extends test_libaom.cc to match the behaviour for other
targets and automatically disable testing for features that are not
available on the machine running the tests.

Change-Id: I7b2d06876bc1d9e410a429536e55247b04a677ef