Fix the sf prune_sub_8x8_partition_level for level 1
In the parent version, when the sf 'prune_sub_8x8_partition_level'
is set to 1, sub_8x8 partitions are pruned using neighbor block
information. But the partitions were pruned even when a top/left
neighbor block was unavailable.

This CL adds a check to disable the pruning if either of the
neighbor blocks is not available. This CL gives bit-exact
output for cpu <= 5.

     Instruction Count        BD-Rate Loss(%)
cpu    Reduction(%)    avg.psnr   ovr.psnr    ssim
 6       -0.042         0.0078     0.0115   -0.0389

STATS_CHANGED

Change-Id: I1d5ef0a8beeae610751d65b685f8cfd46e6b4277