cmake: fix vmaf link w/-DBUILD_SHARED_LIBS=1
Add target_link_libraries(aom_static). This target was added in:
0f23ada000 cmake: Fix aom linkage race condition on MSVC

This target is used to link the examples and tests and was missing the
vmaf library, resulting in undefined symbols.

Also remove -static with -DBUILD_SHARED_LIBS=0 as this may fail if only
the shared object files are available, e.g., in the vmaf package on Arch
Linux.

Bug: aomedia:3464
Fixed: aomedia:3464
Change-Id: I314f961bfaa3819c4474d1a659e1c1575b58206c